% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{Retrieve value labels of labelled data}
\usage{
get_labels(
  x,
  attr.only = FALSE,
  values = NULL,
  non.labelled = FALSE,
  drop.na = TRUE,
  drop.unused = FALSE
)
}
\arguments{
\item{x}{A data frame with variables that have value label attributes (e.g.
from an imported SPSS, SAS or STATA data set, via \code{\link{read_spss}},
\code{\link{read_sas}} or \code{\link{read_stata}}); a variable
(vector) with value label attributes; or a \code{list} of variables
with values label attributes. If \code{x} has no label attributes,
factor levels are returned. See 'Examples'.}

\item{attr.only}{Logical, if \code{TRUE}, labels are only searched for
in the the vector's \code{attributes}; else, if \code{attr.only = FALSE}
and \code{x} has no label attributes, factor levels or string values
are returned. See 'Examples'.}

\item{values}{String, indicating whether the values associated with the
value labels are returned as well. If \code{values = "as.name"}
(or \code{values = "n"}), values are set as \code{names}
attribute of the returned object. If \code{values = "as.prefix"}
(or \code{values = "p"}), values are included as prefix
to each label. See 'Examples'.}

\item{non.labelled}{Logical, if \code{TRUE}, values without labels will
also be included in the returned labels (see \code{\link{fill_labels}}).}

\item{drop.na}{Logical, whether labels of tagged NA values (see \code{\link[haven:tagged_na]{tagged_na()}})
should be included in the return value or not. By default, labelled
(tagged) missing values are not returned. See \code{\link{get_na}}
for more details on tagged NA values.}

\item{drop.unused}{Logical, if \code{TRUE}, unused labels will be removed from
the return value.}
}
\value{
Either a list with all value labels from all variables if \code{x}
          is a \code{data.frame} or \code{list}; a string with the value
          labels, if \code{x} is a variable;
          or \code{NULL} if no value label attribute was found.
}
\description{
This function returns the value labels of labelled data.
}
\examples{
# import SPSS data set
# mydat <- read_spss("my_spss_data.sav")

# retrieve variable labels
# mydat.var <- get_label(mydat)

# retrieve value labels
# mydat.val <- get_labels(mydat)

data(efc)
get_labels(efc$e42dep)

# simple barplot
barplot(table(efc$e42dep))
# get value labels to annotate barplot
barplot(table(efc$e42dep),
        names.arg = get_labels(efc$e42dep),
        main = get_label(efc$e42dep))

# include associated values
get_labels(efc$e42dep, values = "as.name")

# include associated values
get_labels(efc$e42dep, values = "as.prefix")

# get labels from multiple variables
get_labels(list(efc$e42dep, efc$e16sex, efc$e15relat))


# create a dummy factor
f1 <- factor(c("hi", "low", "mid"))
# search for label attributes only
get_labels(f1, attr.only = TRUE)
# search for factor levels as well
get_labels(f1)

# same for character vectors
c1 <- c("higher", "lower", "mid")
# search for label attributes only
get_labels(c1, attr.only = TRUE)
# search for string values as well
get_labels(c1)


# create vector
x <- c(1, 2, 3, 2, 4, NA)
# add less labels than values
x <- set_labels(x, labels = c("yes", "maybe", "no"), force.values = FALSE)
# get labels for labelled values only
get_labels(x)
# get labels for all values
get_labels(x, non.labelled = TRUE)


\dontshow{if (require("haven")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# get labels, including tagged NA values
library(haven)
x <- labelled(c(1:3, tagged_na("a", "c", "z"), 4:1),
              c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
                "Refused" = tagged_na("a"), "Not home" = tagged_na("z")))
# get current NA values
x
get_labels(x, values = "n", drop.na = FALSE)
\dontshow{\}) # examplesIf}


# create vector with unused labels
data(efc)
efc$e42dep <- set_labels(
  efc$e42dep,
  labels = c("independent" = 1, "dependent" = 4, "not used" = 5)
)
get_labels(efc$e42dep)
get_labels(efc$e42dep, drop.unused = TRUE)
get_labels(efc$e42dep, non.labelled = TRUE, drop.unused = TRUE)
}
\seealso{
See vignette \href{../doc/intro_sjlabelled.html}{Labelled Data and the sjlabelled-Package}
           for more details; \code{\link{set_labels}} to manually set value
           labels, \code{\link{get_label}} to get variable labels and
           \code{\link{get_values}} to retrieve the values associated
           with value labels.
}
