% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{set_note}
\alias{set_note}
\alias{get_note}
\title{Deprecated functions}
\usage{
set_note(...)

get_note(...)
}
\arguments{
\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
A list of deprecated functions.
}
