\encoding{UTF-8}
\name{sjlabelled-package}
\alias{sjlabelled-package}
\alias{sjlabelled}
\docType{package}
\title{Labelled Data Utility Functions}
\description{

\strong{Purpose of this package}

Collection of miscellaneous utility functions (especially intended for people coming from other statistical software packages like 'SPSS', and/or who are new to R), supporting following common tasks when working with labelled data:
\itemize{
  \item Reading and writing data between R and other statistical software packages like 'SPSS', 'SAS' or 'Stata'
  \item Easy ways to get, set and change value and variable label attributes, to convert labelled vectors into factors (and vice versa), or to deal with multiple declared missing values etc.
}

}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
