% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_df.R
\name{type_sum}
\alias{type_sum}
\alias{type_sum.label_type_sum}
\title{Provide a succinct summary of an object}
\usage{
type_sum.label_type_sum(x)
}
\arguments{
\item{x}{An object to summarize.}
}
\description{
\code{type_sum.label_type_sum()} is intended for objects of class
   \code{lbl_df} (see \code{\link{lbl_df()}}), to provide information about
   the the variable label when printing tibbles.
}
